Amended the spritesheet_source_frame settings to use ivec2 not vec2

+++ src/Engine.cpp
@@ -40,9 +40,9 @@ void Engine::LoadEngineConfig(void) {
 		windowDimensions = (tileSize * windowGridSize) * windowScaler;
 
 		// Texture Source Frame setup
-		textureBorderSize = glm::vec2(configScript.Get<int>("config.spritesheet_source_frames.texture_border_size.width"), configScript.Get<int>("config.spritesheet_source_frames.texture_border_size.height"));
-		tileTextureFrameSize = glm::vec2(configScript.Get<int>("config.spritesheet_source_frames.tile_frame_dimensions.width"), configScript.Get<int>("config.spritesheet_source_frames.tile_frame_dimensions.height"));
-		entityTextureFrameSize = glm::vec2(configScript.Get<int>("config.spritesheet_source_frames.entity_frame_dimensions.width"), configScript.Get<int>("config.spritesheet_source_frames.entity_frame_dimensions.height"));
+		textureBorderSize = glm::ivec2(configScript.Get<int>("config.spritesheet_source_frames.texture_border_size.width"), configScript.Get<int>("config.spritesheet_source_frames.texture_border_size.height"));
+		tileTextureFrameSize = glm::ivec2(configScript.Get<int>("config.spritesheet_source_frames.tile_frame_dimensions.width"), configScript.Get<int>("config.spritesheet_source_frames.tile_frame_dimensions.height"));
+		entityTextureFrameSize = glm::ivec2(configScript.Get<int>("config.spritesheet_source_frames.entity_frame_dimensions.width"), configScript.Get<int>("config.spritesheet_source_frames.entity_frame_dimensions.height"));
 
 		// Default content setup
 		nameOfTileModel = configScript.Get<std::string>("config.default_content.tile");
