+++ src/Engine.h
@@ -83,9 +83,9 @@ public:
 	Keyboard::Keys keybindMovementRight;
 
 	// Texture frame sizes
-	glm::vec2 textureBorderSize;
-	glm::vec2 tileTextureFrameSize;
-	glm::vec2 entityTextureFrameSize;
+	glm::ivec2 textureBorderSize;
+	glm::ivec2 tileTextureFrameSize;
+	glm::ivec2 entityTextureFrameSize;
 
 	// Constructors
 	Engine(char* gameName);
